chart.js 組み込みに変更

+++ src/db/migrate/20210815053613_create_bodyweights.rb
@@ -2,8 +2,8 @@ class CreateBodyweights < ActiveRecord::Migration[5.2]
   def change
     create_table :bodyweights do |t|
       t.date :start_time
-      t.integer :body_weight
-      t.integer :bodyfat_percentage
+      t.float :body_weight
+      t.float :bodyfat_percentage
       
       t.references :user, foreign_key: true
       t.timestamps
